requirements.txt
Django==1.11.2
django-rest-auth==0.8.1
djangorestframework==3.6.3
Pillow==3.4.1
psycopg2==2.6.2
django-crontab==0.7.1